telegram bot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
environs==9.5.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
marshmallow==3.19.0
moviepy==1.0.3
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
proglog==0.1.10
PyPDF2==3.0.1
python-dotenv==1.0.0
python-telegram-bot==21.6
pytz==2023.3
requests==2.31.0
six==1.16.0
tornado==6.3.2
textblob==0.18.0.post0
tqdm==4.65.0
tzlocal==5.0.1
urllib3==2.0.3
